backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
pydantic==1.10.13
pydantic-settings==2.0.3
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.13.1
sqlalchemy==1.4.41
asyncpg==0.29.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
ruff==0.1.6
black==23.11.0
mypy==1.7.1